feat: permit imputed datasets to have arbitrary sets of files
most imputed datasets come in sets of 22 or 23 files, based on the chromosomes being imputed
separately. if that's the case, that can now be configured for each imputed dataset separately, and
the results will be merged together in the end. individual file's draws are partitioned into
separate subdirectories to prevent any issues with file collisions. more temp() flagging is
eventually needed