feat: use modded bcftools +setGT to randomize minimac genotypes
this currently calls a configured local variant of bcftools, with +setGT modded to allow
randomization of output genotypes according to GP format content. eventually I'd like to have this
pulled from conda, but for the moment relative pathing is fine. the modded variant of bcftools is
available at lightning-auriga/bcftools@setGT_randomize for the time being